+++ AdventOfCode/Inputs/Puzzles/DayFour/DayFourPartOneInput.txt
